Reorganize stateful db p2p resolvers into p2p::standard and p2p::compact
