+++ glue/src/stateful/db/p2p/compact/mod.rs
